+++ apps/admin-ui-domains/src/views/domain/details/domain-mailbox-quota-settings.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { useQueryClient } from '@tanstack/react-query';
-import { Button, Container, CustomHeaderFactory, HoverableRowFactory, Input, ListRow, Padding, Paging, Row, Select, Table, THeader, TrackNumberPerPage, useSnackbar, } from '@zextras/ui-components';
+import { Button, Container, CustomHeaderFactory, HoverableRowFactory, Input, ListRow, Padding, Paging, RouteLeavingGuard, Row, Select, Table, THeader, TrackNumberPerPage, useSnackbar, } from '@zextras/ui-components';
 import { domainByIdKey, flushCache, useIsAdvanced, useUserSettings } from '@zextras/ui-shared';
 import { TFunction } from 'i18next';
 import { isEqual, reduce } from 'lodash-es';
@@ -29,7 +29,6 @@ import { getQuotaUsage } from '../../../services/get-quota-usage-service';
 import { modifyDomain } from '../../../services/modify-domain-service';
 import DownloadCSV from '../../app/shared/download-csv';
 import { MailBoxQuota } from '../../app/types/mailbox_quota';
-import { RouteLeavingGuard } from '../../ui-extras/nav-guard';
 import { BytesToGB, GbToBytes } from '../../utility/utils';
 
 const formatQuota = (quotaUsed: number, quotaLimit: number, t: TFunction): [string, number] => {
@@ -751,15 +750,7 @@ const DomainMailboxQuotaSetting: FC = () => {
         </Row>
       </Container>
 
-      <RouteLeavingGuard when={isDirty} onSave={onSave}>
-        <ds-text as="p">
-          {t(
-            'label.unsaved_changes_line1',
-            'Are you sure you want to leave this page without saving?',
-          )}
-        </ds-text>
-        <ds-text as="p">{t('label.unsaved_changes_line2', 'All your unsaved changes will be lost')}</ds-text>
-      </RouteLeavingGuard>
+      <RouteLeavingGuard when={isDirty} onSave={onSave} />
     </Container>
   );
 };